test(source-qa): relocate website-crawl case to integration; ALLOWED_SERVE now empty (#664)
The website-source update test crawls the network deep inside
ensureWebsiteMirror -> crawlWebsite -> fetchWebsitePage -> fetchWithRetry,
behind the generic source-provider sync() boundary that takes no injectable
fetch. Per the seam design this is a genuinely transport-shaped case, kept as
a thin integration test against a real Bun.serve rather than de-socketed.
Moved to tests/integration/website-source-update.test.ts using the sanctioned
withIsolatedAkmStorage helper; the remaining source-qa-fixes assertions
(kind/writable/name, plus the spyOn'd git-update case) stay pure in the unit
tier.

ALLOWED_SERVE is now empty — every unit-tier Bun.serve has been migrated or
relocated. UNIT_PURITY_BASELINE 18 -> 17.
