+++ scripts/lint-tests-unit-purity.ts
@@ -38,8 +38,14 @@ const repoRoot = path.resolve(import.meta.dir, "..");
 
 // ── Grandfather allowlists (shrink-only) ─────────────────────────────────────
 
-/** Unit-tier files that currently stand up `Bun.serve`. Shrink-only. */
-const ALLOWED_SERVE = new Set<string>(["tests/source-qa-fixes.test.ts"]);
+/**
+ * Unit-tier files that currently stand up `Bun.serve`. Shrink-only.
+ *
+ * EMPTY — every unit-tier `Bun.serve` has been migrated onto an injected
+ * `HttpClient` seam (#664 Seam 1) or relocated to `tests/integration/`. A NEW
+ * `Bun.serve` in a unit file now fails the lint with no grandfather escape.
+ */
+const ALLOWED_SERVE = new Set<string>([]);
 
 /** Unit-tier files that currently spawn a real subprocess. Shrink-only. */
 const ALLOWED_SPAWN = new Set<string>([
@@ -67,7 +73,7 @@ const ALLOWED_SPAWN = new Set<string>([
  * change whenever you remove an entry; never raise it. Meta-test:
  * `tests/lint-unit-purity-ratchet.test.ts`.
  */
-export const UNIT_PURITY_BASELINE = 18;
+export const UNIT_PURITY_BASELINE = 17;
 
 export function combinedAllowlistSize(): number {
   return ALLOWED_SERVE.size + ALLOWED_SPAWN.size;
